refactored types in redux store
