+++ src/Pages/MyPage.tsx
@@ -3,7 +3,7 @@ import { useHistory } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { selectUser } from "../Store/users/selector";
 import { selectAppLoading } from "../Store/appState/selector";
-import { Trip } from "../Types/model";
+import { Trip } from "../Types/tripTypes";
 import TripCard from "../Components/TripCard";
 import NewTripModal from "../Components/NewTripModal";
 import Loading from "../Components/AppStateComponents/Loading";
@@ -28,8 +28,10 @@ export default function MyPage() {
 
     //@ts-ignore
     const about = user.about
-      .split("\n")
-      .map((paragraph, i) => <Text key={i}>{paragraph}</Text>);
+      ? user.about
+          .split("\n")
+          .map((paragraph, i) => <Text key={i}>{paragraph}</Text>)
+      : [];
 
     return (
       <Box w="100vw" m="auto" bg="blackAlpha.300" p="2vh" top="1" mt="3rem">
