+++ src/Pages/TripDetails.tsx
@@ -4,11 +4,12 @@ import { opencageAPIkey } from "../config/constants";
 import { useSelector, useDispatch } from "react-redux";
 import { useParams } from "react-router-dom";
 import { showMessageWithTimeout } from "../Store/appState/actions";
-import { fetchSpecificTrip, endTrip } from "../Store/trips/actions";
+import { fetchSpecificTrip } from "../Store/trips/actions";
+import { endTrip } from "../Store/users/actions";
 import { selectIdTrip } from "../Store/trips/selector";
 import { selectUser, selectUserTrip } from "../Store/users/selector";
 import { selectAppLoading } from "../Store/appState/selector";
-import { Trip, Post, DefaultMarker } from "../Types/model";
+import { Trip, Post, DefaultMarker } from "../Types/tripTypes";
 import Slider from "../Components/slider";
 import GoogleMaps from "../Components/GoogleMaps";
 import NewPostModal from "../Components/NewPostModal";
