+++ src/Types/appStateTypes.ts
@@ -1,0 +1,10 @@
+export type AppState = {
+  loading: Boolean;
+  message: Message | null;
+};
+
+export type Message = {
+  variant: string;
+  dismissable: boolean;
+  text: string;
+};
