+++ src/config/constants.tsx
@@ -1,4 +1,5 @@
-export const apiUrl = process.env.REACT_APP_API_URL || "localhost:5000";
+// export const apiUrl = process.env.REACT_APP_API_URL || "localhost:5000";
+export const apiUrl = "http://localhost:5000";
 export const opencageAPIkey = process.env.REACT_APP_OPENCAGE_API_KEY;
 export const googleAPIkey = process.env.REACT_APP_GOOGLE_API;
 export const cloudinaryUrl = process.env.REACT_APP_CLOUDINARY_URL;
